add authorization in asset crud.

+++ app/Http/Controllers/AssetController.php
@@ -12,22 +12,26 @@ class AssetController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
+     * @throws
      */
     public function index()
     {
-        $assets = Asset::all();
-        return response()->json($assets);
+        $this->authorize('view-any',Asset::class);
+        $assets = Asset::paginate(10);
+        return response()->json($assets,200);
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param AssetStoreRequest $request
+     * @return Response
+     * @throws
      */
     public function store(AssetStoreRequest $request)
     {
+        $this->authorize('create', Asset::class);
         $validated = $request->validated();
         $asset = Asset::create($validated);
         return response()->json($asset, Response::HTTP_CREATED);
@@ -36,24 +40,27 @@ class AssetController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $asset
-     * @return \Illuminate\Http\Response
+     * @param Asset $asset
+     * @return Response
+     * @throws
      */
-    public function show($asset)
+    public function show(Asset $asset)
     {
-        return response()->json(Asset::findOrFail($asset));
+        $this->authorize('view', $asset);
+        return response()->json($asset,200);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $asset
-     * @return \Illuminate\Http\Response
+     * @param AssetStoreRequest $request
+     * @param Asset $asset
+     * @return Response
+     * @throws
      */
-    public function update(AssetStoreRequest $request, $asset)
+    public function update(AssetStoreRequest $request, Asset $asset)
     {
-        $asset = Asset::findOrFail($asset);
+        $this->authorize('update', $asset);
         $validated = $request->validated();
         $asset->update($validated);
         return response()->json($asset, Response::HTTP_OK);
@@ -62,12 +69,13 @@ class AssetController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $asset
-     * @return \Illuminate\Http\Response
+     * @param Asset $asset
+     * @return Response
+     * @throws
      */
-    public function destroy($asset)
+    public function destroy(Asset $asset)
     {
-        $asset = Asset::findOrFail($asset);
+        $this->authorize('delete', $asset);
         $asset->delete();
         return response()->json(null, Response::HTTP_NO_CONTENT);
     }
